Coerce passed type to TS before invoking ::name()
If we're running the macro on a type which has an in-scope trait
containing a `::name` member, the compiler will not know which to
invoke.

This clarifies that in the macro's context, we want to invoke the
TS::name trait method by first casting the argument to the TS trait.

This pattern is in some other parts of the macro, I think this one was
just triggered previously.

Signed-off-by: Paul R. Tagliamonte <paultag@gmail.com>