opt out of automatic export of bindings
this patch introduces a new feature - "export-bindings" - to grant
downstream more control over when the bindings are generated.

With the feature disabled in Cargo.toml, the bindings are not exported
automatically on every test run. The user can choose when to export by
explicitly enabling the feature with the --features cargo test flag.

For example, in my Cargo.toml:

```toml
[dependencies]
ts-rs = { version = "11", default-features = false, features = [
  "serde-compat",
  "no-serde-warnings",
] }
```

And at the time I need to export the bindings:

    cargo test --features export-bindings -- 'export_bindings_'

This is somewhat related to #62 and #253.